ARMOA005- 1771 Cambio en tipo de dato para valores cantidades de nominacion.

+++ Molinos.Scato.Database/dbo/Tables/NominacionDatoTecnicoDestino.sql
@@ -1,7 +1,7 @@
 create table NominacionDatoTecnicoDestino(
 Id int  IDENTITY (1, 1) NOT NULL,
 Destino_Id int  not null,
-Cantidad int ,
+Cantidad DECIMAL(20, 3) ,
 NominacionDatoTecnico_Id int not null,
 CONSTRAINT [PK_dbo.NominacionDatoTecnicoDestino] PRIMARY KEY CLUSTERED ([Id] ASC),
 CONSTRAINT [FK_dbo.NominacionDatoTecnicoDestino_dbo.Destino_Destino_Id] FOREIGN KEY ([Destino_Id]) REFERENCES [dbo].[Destino] ([Id]),
